docs: fix some stackblitz examples (#216)
