+++ projects/ng-aquila/documentation/assets/stack-blitz/angular.json
@@ -25,8 +25,10 @@
             "tsConfig": "tsconfig.app.json",
             "assets": [],
             "styles": [
+              "node_modules/@angular/cdk/overlay-prebuilt.css",
               "src/styles.scss",
               "node_modules/@aposin/ng-aquila/css/normalize.css",
+              "node_modules/@aposin/ng-aquila/css/utilities.css",
               "node_modules/@aposin/ng-aquila/themes/aposin.css"
             ],
             "scripts": []
