+++ projects/ng-aquila/documentation/examples/accessibility/accessibility-high-contrast-svg/accessibility-high-contrast-svg-example.ts
@@ -4,6 +4,7 @@ import { Component } from '@angular/core';
 * @title High contrast svg example
 */
 @Component({
+  selector: 'accessibility-high-contrast-svg-example',
   templateUrl: './accessibility-high-contrast-svg-example.html',
   styleUrls: ['./accessibility-high-contrast-svg-example.css']
 })
