+++ projects/ng-aquila/documentation/examples/error/error-custom-matcher-formfield/error-custom-matcher-formfield-example.html
@@ -1,9 +1,9 @@
-  <nx-formfield nxLabel="Label">
-    <input nxInput [formControl]="emailFormControl"/>
-    <span nxFormfieldHint>
-      This field will show an error while typing.
-    </span>
-    <nx-error nxFormfieldError>
-     Please enter a valid email address.
-    </nx-error>
-  </nx-formfield>
+<nx-formfield nxLabel="Label">
+  <input nxInput [formControl]="emailFormControl"/>
+  <span nxFormfieldHint>
+    This field will show an error while typing.
+  </span>
+  <nx-error nxFormfieldError>
+    Please enter a valid email address.
+  </nx-error>
+</nx-formfield>
