+++ projects/ng-aquila/documentation/examples/margin/margin-sizes/margin-sizes-example.css
@@ -1,3 +1,7 @@
+.margin-box {
+  background: var(--documentation-margin-example-background-color, #1c7997);
+}
+
 @media screen and (-ms-high-contrast: active) {
   .margin-box {
     background-color: windowText;
