+++ projects/ng-aquila/documentation/examples/margin/margin-usage/margin-usage-example.css
@@ -1,19 +1,22 @@
 .wrapper {
+  background: var(--documentation-margin-example-background-color, #1c7997);
   display: table;
   align-items: center;
   height: 100%;
   min-height: 100%;
 }
 
-.margin-box {
-  width: 400px;
-  display: block;
-  line-height: 20px;
-  padding: 8px;
-}
-
 @media screen and (-ms-high-contrast: active) {
   .wrapper {
     background-color: windowText;
   }
+}
+
+.margin-box {
+  border: 1px solid var(--documentation-margin-example-background-color, #1c7997);
+  background-color: var(--ui-01, #0f2a3d);
+  display: block;
+  line-height: 20px;
+  padding: 8px;
+  width: 400px;
 }
